+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneAccountProtectionLocalAdministratorPasswordSolutionPolicy/MSFT_IntuneAccountProtectionLocalAdministratorPasswordSolutionPolicy.psm1
@@ -425,7 +425,7 @@ function Set-TargetResource
         }
         $newPolicy = New-MgBetaDeviceManagementConfigurationPolicy -bodyParameter $createParameters
 
-        $assignmentsHash = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignments
+        $assignmentsHash = ConvertTo-IntunePolicyAssignment -IncludeDeviceFilter:$true -Assignments $Assignments
         Update-DeviceConfigurationPolicyAssignment `
             -DeviceConfigurationPolicyId $newPolicy.Id `
             -Targets $assignmentsHash
@@ -449,7 +449,7 @@ function Set-TargetResource
             -Settings $settings
 
         #region update policy assignments
-        $assignmentsHash = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignments
+        $assignmentsHash = ConvertTo-IntunePolicyAssignment -IncludeDeviceFilter:$true -Assignments $Assignments
         Update-DeviceConfigurationPolicyAssignment `
             -DeviceConfigurationPolicyId $currentPolicy.Identity `
             -Targets $assignmentsHash
@@ -1159,7 +1159,7 @@ function Update-DeviceManagementConfigurationPolicy
     Invoke-MgGraphRequest -Method PUT `
         -Uri $Uri `
         -ContentType 'application/json' `
-        -Body ($policy | ConvertTo-Json -Depth 20) 4> out-null
+        -Body ($policy | ConvertTo-Json -Depth 20) 4> $null
 }
 
 function Get-DeviceManagementConfigurationPolicyAssignment
@@ -1177,7 +1177,7 @@ function Get-DeviceManagementConfigurationPolicyAssignment
         $configurationPolicyAssignments = @()
 
         $Uri = "https://graph.microsoft.com/beta/deviceManagement/configurationPolicies/$DeviceManagementConfigurationPolicyId/assignments"
-        $results = Invoke-MgGraphRequest -Method GET -Uri $Uri -ErrorAction Stop 4> out-null
+        $results = Invoke-MgGraphRequest -Method GET -Uri $Uri -ErrorAction Stop 4> $null
         foreach ($result in $results.value.target)
         {
             $configurationPolicyAssignments += @{
@@ -1192,7 +1192,7 @@ function Get-DeviceManagementConfigurationPolicyAssignment
         while ($results.'@odata.nextLink')
         {
             $Uri = $results.'@odata.nextLink'
-            $results = Invoke-MgGraphRequest -Method GET -Uri $Uri -ErrorAction Stop 4> out-null
+            $results = Invoke-MgGraphRequest -Method GET -Uri $Uri -ErrorAction Stop 4> $null
             foreach ($result in $results.value.target)
             {
                 $configurationPolicyAssignments += @{